move mod into specific file
